add content .. question

+++ src/main/java/com/example/demo/controller/publishController.java
@@ -17,13 +17,11 @@ import javax.servlet.http.HttpServletRequest;
 
 @Controller
 public class publishController {
-
-
     @Autowired
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -43,7 +41,7 @@ public class publishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title", title);
@@ -61,14 +59,11 @@ public class publishController {
             model.addAttribute("error", "标签不能为空");
             return "publish";
         }
-
         User user = (User) request.getSession().getAttribute("user");
         if (user == null) {
             model.addAttribute("error", "用户未登录");
             return "publish";
         }
-
-
         Question question = new Question();
         question.setTitle(title);
         question.setDescription(description);
